refactor: drop deno, and sadly, tests
