+++ .vscode/settings.json
@@ -1,8 +1,9 @@
 {
-  // "deno.enable": true,
-  "deno.enable": false,
+  "deno.enable": true,
+  // "deno.enable": false,
   "deno.codeLens.test": true,
   "deno.importMap": "./import_map.json",
+  "deno.unstable": true,
   "editor.defaultFormatter": "esbenp.prettier-vscode",
   "editor.formatOnSave": true,
   "git.ignoreLimitWarning": true
