+++ package.json
@@ -15,6 +15,7 @@
   "devDependencies": {
     "ava": "4.1.0",
     "browser-env": "3.3.0",
+    "esbuild": "0.17.7",
     "prettier": "^2.7.1",
     "prettier-plugin-organize-imports": "3.2.2",
     "ts-node": "^10.9.1",
@@ -36,5 +37,8 @@
   "workspaces": [
     "packages/*",
     "apps/*"
-  ]
+  ],
+  "dependencies": {
+    "date-fns": "2.28.0"
+  }
 }
