+++ packages/cleanair-sensor-airmapgps/src/download/mod_web.ts
@@ -1,0 +1,5 @@
+import { closestTo } from "date-fns";
+import { createModule } from "./mod";
+
+export const moduleResources = { closestTo };
+export default createModule(moduleResources);
