+++ packages/cleanair-sensor-common/mod.ts
@@ -1,4 +1,4 @@
-export * from "./mapbox.ts";
+export * from "./mapbox";
 export { type GeoJSON } from "https://deno.land/x/d3_4_deno@v6.2.0.9/src/d3-geo/geojson.js?s=GeoJSON";
 
 export type MapGetLevels<P = unknown> = ({
