+++ packages/cleanair-sensor-flow/src/__test__/parsers/flow.deno.test.ts.bak
@@ -1,5 +1,5 @@
-import mod, { moduleResources } from "../../download/mod_deno.ts";
-import { assertEquals } from "https://deno.land/std@0.129.0/testing/asserts.ts";
+import mod, { moduleResources } from "../../download/mod_deno";
+import { assertEquals } from "https://deno.land/std@0.129.0/testing/asserts";
 
 const measures = [
   {
