--- packages/cleanair-sensor-flow/src/download/mod_node.ts
@@ -1,1 +1,0 @@
-// app must pass in resources, see mod.ts
