+++ packages/cleanair-sensor-miniwras/mod.ts
@@ -1,5 +1,5 @@
-export * from "./src/download/mod_deno.ts";
-export { default } from "./src/download/mod_deno.ts";
-export * from "./src/interfaces.ts";
+export * from "./src/download/mod";
+export * from "./src/interfaces";
+export { default } from "./src/download/mod_web";
 // mapbox, hook are not deno compliant
-// export * as mapbox from "./src/mapbox.ts";
+// export * as mapbox from "./src/mapbox";
