--- packages/cleanair-sensor-miniwras/src/download/mod_deno.ts
@@ -1,3 +1,0 @@
-import closestTo from "https://deno.land/x/date_fns@v2.22.1/closestTo/index.js";
-import { createModule } from "./mod.ts";
-export default createModule({ closestTo });
