+++ packages/cleanair-sensor-miniwras/src/download/mod_web.ts
@@ -1,0 +1,3 @@
+import { closestTo } from "date-fns";
+import { createModule } from "./mod";
+export default createModule({ closestTo });
