+++ packages/cleanair-sensor-pocketlabs/mod.ts
@@ -1,5 +1,5 @@
-export * from "./src/download.ts";
-export * from "./src/streams/parse-pocketlabs-stream.ts";
+export * from "./src/download";
+export * from "./src/streams/parse-pocketlabs-stream";
 // hook, mapbox are not deno compliant. keep it out!
-// export * as mapbox from "./src/mapbox.ts";
-// export * from "./src/hook.ts";
+// export * as mapbox from "./src/mapbox";
+// export * from "./src/hook";
