+++ packages/cleanair-sensor-pocketlabs/src/resources.ts
@@ -1,7 +1,7 @@
-import * as download from "./download.ts";
-import * as mapbox from "./mapbox.ts";
+import * as download from "./download";
+import * as mapbox from "./mapbox";
 import * as stream from "./streams/parse-pocketlabs-stream";
-export * from "./download.ts";
+export * from "./download";
 export const getResources = () => {
   return { download, mapbox, stream };
 };
