+++ packages/cleanair-sensor-strava-gpx/__test__/parse.test.ts.bak
@@ -1,5 +1,5 @@
-import { assertEquals } from "https://deno.land/std@0.129.0/testing/asserts.ts";
-import * as mod from "../mod.ts";
+import { assertEquals } from "https://deno.land/std@0.129.0/testing/asserts";
+import * as mod from "../mod";
 import { DOMParser } from "https://esm.sh/linkedom";
 
 // deno-lint-ignore no-explicit-any
